separate out from click_test.go things that seems to obviously belong to other impl file companion test files, create some of those as necessary
