+++ snappy/click_test.go
@@ -20,20 +20,16 @@
 package snappy
 
 import (
-	"fmt"
 	"io/ioutil"
 	"os"
 	"path/filepath"
-	"regexp"
 	"strings"
 
 	. "gopkg.in/check.v1"
 
-	"github.com/ubuntu-core/snappy/arch"
 	"github.com/ubuntu-core/snappy/dirs"
 	"github.com/ubuntu-core/snappy/osutil"
 	"github.com/ubuntu-core/snappy/systemd"
-	"github.com/ubuntu-core/snappy/timeout"
 )
 
 func (s *SnapTestSuite) TestLocalSnapInstall(c *C) string {
@@ -351,73 +347,6 @@ func (s *SnapTestSuite) TestClickCopyDataNoUserHomes(c *C) {
 	c.Assert(err, IsNil)
 }
 
-const expectedWrapper = `#!/bin/sh
-set -e
-
-# snap info (deprecated)
-export SNAP_APP_PATH="/snaps/pastebinit/1.4.0.0.1/"
-export SNAP_APP_DATA_PATH="/var/lib/snaps/pastebinit/1.4.0.0.1/"
-export SNAP_APP_USER_DATA_PATH="$HOME/snaps/pastebinit/1.4.0.0.1/"
-
-# snap info
-export SNAP="/snaps/pastebinit/1.4.0.0.1/"
-export SNAP_DATA="/var/lib/snaps/pastebinit/1.4.0.0.1/"
-export SNAP_NAME="pastebinit"
-export SNAP_VERSION="1.4.0.0.1"
-export SNAP_ARCH="%[1]s"
-export SNAP_USER_DATA="$HOME/snaps/pastebinit/1.4.0.0.1/"
-
-if [ ! -d "$SNAP_USER_DATA" ]; then
-   mkdir -p "$SNAP_USER_DATA"
-fi
-export HOME="$SNAP_USER_DATA"
-
-# Snap name is: pastebinit
-# App name is: pastebinit
-
-ubuntu-core-launcher pastebinit.pastebinit pastebinit_pastebinit_1.4.0.0.1 /snaps/pastebinit/1.4.0.0.1/bin/pastebinit "$@"
-`
-
-func (s *SnapTestSuite) TestSnappyGenerateSnapBinaryWrapper(c *C) {
-	binary := &AppYaml{Name: "pastebinit", Command: "bin/pastebinit"}
-	pkgPath := "/snaps/pastebinit/1.4.0.0.1/"
-	aaProfile := "pastebinit_pastebinit_1.4.0.0.1"
-	m := snapYaml{Name: "pastebinit",
-		Version: "1.4.0.0.1"}
-
-	expected := fmt.Sprintf(expectedWrapper, arch.UbuntuArchitecture())
-
-	generatedWrapper, err := generateSnapBinaryWrapper(binary, pkgPath, aaProfile, &m)
-	c.Assert(err, IsNil)
-	c.Assert(generatedWrapper, Equals, expected)
-}
-
-func (s *SnapTestSuite) TestSnappyGenerateSnapBinaryWrapperIllegalChars(c *C) {
-	binary := &AppYaml{Name: "bin/pastebinit\nSomething nasty"}
-	pkgPath := "/snaps/pastebinit.mvo/1.4.0.0.1/"
-	aaProfile := "pastebinit.mvo_pastebinit_1.4.0.0.1"
-	m := snapYaml{Name: "pastebinit",
-		Version: "1.4.0.0.1"}
-
-	_, err := generateSnapBinaryWrapper(binary, pkgPath, aaProfile, &m)
-	c.Assert(err, NotNil)
-}
-
-func (s *SnapTestSuite) TestSnappyBinPathForBinaryNoExec(c *C) {
-	binary := &AppYaml{Name: "pastebinit", Command: "bin/pastebinit"}
-	pkgPath := "/snaps/pastebinit.mvo/1.0/"
-	c.Assert(binPathForBinary(pkgPath, binary), Equals, "/snaps/pastebinit.mvo/1.0/bin/pastebinit")
-}
-
-func (s *SnapTestSuite) TestSnappyBinPathForBinaryWithExec(c *C) {
-	binary := &AppYaml{
-		Name:    "pastebinit",
-		Command: "bin/random-pastebin",
-	}
-	pkgPath := "/snaps/pastebinit.mvo/1.1/"
-	c.Assert(binPathForBinary(pkgPath, binary), Equals, "/snaps/pastebinit.mvo/1.1/bin/random-pastebin")
-}
-
 func (s *SnapTestSuite) TestSnappyHandleBinariesOnUpgrade(c *C) {
 	snapYamlContent := `name: foo
 apps:
@@ -495,342 +424,6 @@ apps:
 	c.Assert(err, IsNil)
 
 	c.Assert(allSystemctl, HasLen, 0)
-
-}
-
-func (s *SnapTestSuite) TestAddPackageServicesStripsGlobalRootdir(c *C) {
-	// ensure that even with a global rootdir the paths in the generated
-	// .services file are setup correctly (i.e. that the global root
-	// is stripped)
-	c.Assert(dirs.GlobalRootDir, Not(Equals), "/")
-
-	yamlFile, err := makeInstalledMockSnap(s.tempdir, "")
-	c.Assert(err, IsNil)
-	m, err := parseSnapYamlFile(yamlFile)
-	c.Assert(err, IsNil)
-	baseDir := filepath.Dir(filepath.Dir(yamlFile))
-	err = addPackageServices(m, baseDir, false, nil)
-	c.Assert(err, IsNil)
-
-	content, err := ioutil.ReadFile(filepath.Join(s.tempdir, "/etc/systemd/system/hello-snap_svc1_1.10.service"))
-	c.Assert(err, IsNil)
-
-	baseDirWithoutRootPrefix := "/snaps/" + helloSnapComposedName + "/1.10"
-	verbs := []string{"Start", "Stop", "StopPost"}
-	bins := []string{"hello", "goodbye", "missya"}
-	for i := range verbs {
-		expected := fmt.Sprintf("Exec%s=/usr/bin/ubuntu-core-launcher hello-snap.svc1 %s_svc1_1.10 %s/bin/%s", verbs[i], helloSnapComposedName, baseDirWithoutRootPrefix, bins[i])
-		c.Check(string(content), Matches, "(?ms).*^"+regexp.QuoteMeta(expected)) // check.v1 adds ^ and $ around the regexp provided
-	}
-}
-
-func (s *SnapTestSuite) TestAddPackageBinariesStripsGlobalRootdir(c *C) {
-	// ensure that even with a global rootdir the paths in the generated
-	// .services file are setup correctly (i.e. that the global root
-	// is stripped)
-	c.Assert(dirs.GlobalRootDir, Not(Equals), "/")
-
-	yamlFile, err := makeInstalledMockSnap(s.tempdir, "")
-	c.Assert(err, IsNil)
-	m, err := parseSnapYamlFile(yamlFile)
-	c.Assert(err, IsNil)
-	baseDir := filepath.Dir(filepath.Dir(yamlFile))
-	err = addPackageBinaries(m, baseDir)
-	c.Assert(err, IsNil)
-
-	content, err := ioutil.ReadFile(filepath.Join(s.tempdir, "/snaps/bin/hello-snap.hello"))
-	c.Assert(err, IsNil)
-
-	needle := `
-ubuntu-core-launcher hello-snap.hello hello-snap_hello_1.10 /snaps/hello-snap/1.10/bin/hello "$@"
-`
-	c.Assert(string(content), Matches, "(?ms).*"+regexp.QuoteMeta(needle)+".*")
-}
-
-var (
-	expectedServiceWrapperFmt = `[Unit]
-Description=A fun webserver
-%s
-X-Snappy=yes
-
-[Service]
-ExecStart=/usr/bin/ubuntu-core-launcher xkcd-webserver.xkcd-webserver xkcd-webserver_xkcd-webserver_0.3.4 /snaps/xkcd-webserver/0.3.4/bin/foo start
-Restart=on-failure
-WorkingDirectory=/var/lib/snaps/xkcd-webserver/0.3.4/
-Environment="SNAP_APP=xkcd-webserver_xkcd-webserver_0.3.4" "SNAP=/snaps/xkcd-webserver/0.3.4/" "SNAP_DATA=/var/lib/snaps/xkcd-webserver/0.3.4/" "SNAP_NAME=xkcd-webserver" "SNAP_VERSION=0.3.4" "SNAP_ARCH=%[3]s" "SNAP_USER_DATA=/root/snaps/xkcd-webserver/0.3.4/" "SNAP_APP_PATH=/snaps/xkcd-webserver/0.3.4/" "SNAP_APP_DATA_PATH=/var/lib/snaps/xkcd-webserver/0.3.4/" "SNAP_APP_USER_DATA_PATH=/root/snaps/xkcd-webserver/0.3.4/"
-ExecStop=/usr/bin/ubuntu-core-launcher xkcd-webserver.xkcd-webserver xkcd-webserver_xkcd-webserver_0.3.4 /snaps/xkcd-webserver/0.3.4/bin/foo stop
-ExecStopPost=/usr/bin/ubuntu-core-launcher xkcd-webserver.xkcd-webserver xkcd-webserver_xkcd-webserver_0.3.4 /snaps/xkcd-webserver/0.3.4/bin/foo post-stop
-TimeoutStopSec=30
-%[2]s
-
-[Install]
-WantedBy=multi-user.target
-`
-	expectedServiceAppWrapper     = fmt.Sprintf(expectedServiceWrapperFmt, "After=ubuntu-snappy.frameworks.target\nRequires=ubuntu-snappy.frameworks.target", "Type=simple\n", arch.UbuntuArchitecture())
-	expectedServiceFmkWrapper     = fmt.Sprintf(expectedServiceWrapperFmt, "Before=ubuntu-snappy.frameworks.target\nAfter=ubuntu-snappy.frameworks-pre.target\nRequires=ubuntu-snappy.frameworks-pre.target", "Type=dbus\nBusName=foo.bar.baz", arch.UbuntuArchitecture())
-	expectedSocketUsingWrapper    = fmt.Sprintf(expectedServiceWrapperFmt, "After=ubuntu-snappy.frameworks.target xkcd-webserver_xkcd-webserver_0.3.4.socket\nRequires=ubuntu-snappy.frameworks.target xkcd-webserver_xkcd-webserver_0.3.4.socket", "Type=simple\n", arch.UbuntuArchitecture())
-	expectedTypeForkingFmkWrapper = fmt.Sprintf(expectedServiceWrapperFmt, "After=ubuntu-snappy.frameworks.target\nRequires=ubuntu-snappy.frameworks.target", "Type=forking\n", arch.UbuntuArchitecture())
-)
-
-func (s *SnapTestSuite) TestSnappyGenerateSnapServiceTypeForking(c *C) {
-	service := &AppYaml{
-		Name:        "xkcd-webserver",
-		Command:     "bin/foo start",
-		Stop:        "bin/foo stop",
-		PostStop:    "bin/foo post-stop",
-		StopTimeout: timeout.DefaultTimeout,
-		Description: "A fun webserver",
-		Daemon:      "forking",
-	}
-	pkgPath := "/snaps/xkcd-webserver/0.3.4/"
-	aaProfile := "xkcd-webserver_xkcd-webserver_0.3.4"
-	m := snapYaml{Name: "xkcd-webserver",
-		Version: "0.3.4"}
-
-	generatedWrapper, err := generateSnapServicesFile(service, pkgPath, aaProfile, &m)
-	c.Assert(err, IsNil)
-	c.Assert(generatedWrapper, Equals, expectedTypeForkingFmkWrapper)
-}
-
-func (s *SnapTestSuite) TestSnappyGenerateSnapServiceAppWrapper(c *C) {
-	service := &AppYaml{
-		Name:        "xkcd-webserver",
-		Command:     "bin/foo start",
-		Stop:        "bin/foo stop",
-		PostStop:    "bin/foo post-stop",
-		StopTimeout: timeout.DefaultTimeout,
-		Description: "A fun webserver",
-		Daemon:      "simple",
-	}
-	pkgPath := "/snaps/xkcd-webserver/0.3.4/"
-	aaProfile := "xkcd-webserver_xkcd-webserver_0.3.4"
-	m := snapYaml{Name: "xkcd-webserver",
-		Version: "0.3.4"}
-
-	generatedWrapper, err := generateSnapServicesFile(service, pkgPath, aaProfile, &m)
-	c.Assert(err, IsNil)
-	c.Assert(generatedWrapper, Equals, expectedServiceAppWrapper)
-}
-
-func (s *SnapTestSuite) TestSnappyGenerateSnapServiceRestart(c *C) {
-	service := &AppYaml{
-		Name:        "xkcd-webserver",
-		RestartCond: systemd.RestartOnAbort,
-		Daemon:      "simple",
-	}
-	pkgPath := "/snaps/xkcd-webserver/0.3.4/"
-	aaProfile := "xkcd-webserver_xkcd-webserver_0.3.4"
-	m := snapYaml{
-		Name:    "xkcd-webserver",
-		Version: "0.3.4",
-	}
-
-	generatedWrapper, err := generateSnapServicesFile(service, pkgPath, aaProfile, &m)
-	c.Assert(err, IsNil)
-	c.Assert(generatedWrapper, Matches, `(?ms).*^Restart=on-abort$.*`)
-}
-
-func (s *SnapTestSuite) TestSnappyGenerateSnapServiceWrapperWhitelist(c *C) {
-	service := &AppYaml{Name: "xkcd-webserver",
-		Command:     "bin/foo start",
-		Stop:        "bin/foo stop",
-		PostStop:    "bin/foo post-stop",
-		StopTimeout: timeout.DefaultTimeout,
-		Description: "A fun webserver\nExec=foo",
-		Daemon:      "simple",
-	}
-	pkgPath := "/snaps/xkcd-webserver.canonical/0.3.4/"
-	aaProfile := "xkcd-webserver.canonical_xkcd-webserver_0.3.4"
-	m := snapYaml{Name: "xkcd-webserver",
-		Version: "0.3.4"}
-
-	_, err := generateSnapServicesFile(service, pkgPath, aaProfile, &m)
-	c.Assert(err, NotNil)
-}
-
-func (s *SnapTestSuite) TestServiceWhitelistSimple(c *C) {
-	c.Assert(verifyAppYaml(&AppYaml{Name: "foo"}), IsNil)
-	c.Assert(verifyAppYaml(&AppYaml{Description: "foo"}), IsNil)
-	c.Assert(verifyAppYaml(&AppYaml{Command: "foo"}), IsNil)
-	c.Assert(verifyAppYaml(&AppYaml{Stop: "foo"}), IsNil)
-	c.Assert(verifyAppYaml(&AppYaml{PostStop: "foo"}), IsNil)
-}
-
-func (s *SnapTestSuite) TestServiceWhitelistIllegal(c *C) {
-	c.Assert(verifyAppYaml(&AppYaml{Name: "x\n"}), NotNil)
-	c.Assert(verifyAppYaml(&AppYaml{Description: "foo\n"}), NotNil)
-	c.Assert(verifyAppYaml(&AppYaml{Command: "foo\n"}), NotNil)
-	c.Assert(verifyAppYaml(&AppYaml{Stop: "foo\n"}), NotNil)
-	c.Assert(verifyAppYaml(&AppYaml{PostStop: "foo\n"}), NotNil)
-}
-
-func (s *SnapTestSuite) TestVerifyAppDaemonValue(c *C) {
-	c.Assert(verifyAppYaml(&AppYaml{Daemon: "oneshot"}), IsNil)
-	c.Assert(verifyAppYaml(&AppYaml{Daemon: "nono"}), ErrorMatches, `"daemon" field contains invalid value "nono"`)
-}
-
-func (s *SnapTestSuite) TestServiceWhitelistError(c *C) {
-	err := verifyAppYaml(&AppYaml{Name: "x\n"})
-	c.Assert(err.Error(), Equals, `app description field 'Name' contains illegal "x\n" (legal: '^[A-Za-z0-9/. _#:-]*$')`)
-}
-
-func (s *SnapTestSuite) TestBinariesWhitelistSimple(c *C) {
-	c.Assert(verifyAppYaml(&AppYaml{Name: "foo"}), IsNil)
-	c.Assert(verifyAppYaml(&AppYaml{Command: "foo"}), IsNil)
-}
-
-func (s *SnapTestSuite) TestUsesWhitelistSimple(c *C) {
-	c.Check(verifyPlugYaml(&plugYaml{
-		Interface: "old-security",
-		SecurityDefinitions: SecurityDefinitions{
-			SecurityTemplate: "foo"},
-	}), IsNil)
-	c.Check(verifyPlugYaml(&plugYaml{
-		Interface: "old-security",
-		SecurityDefinitions: SecurityDefinitions{
-			SecurityPolicy: &SecurityPolicyDefinition{
-				AppArmor: "foo"},
-		},
-	}), IsNil)
-}
-
-func (s *SnapTestSuite) TestBinariesWhitelistIllegal(c *C) {
-	c.Assert(verifyAppYaml(&AppYaml{Name: "test!me"}), NotNil)
-	c.Assert(verifyAppYaml(&AppYaml{Name: "x\n"}), NotNil)
-	c.Assert(verifyAppYaml(&AppYaml{Command: "x\n"}), NotNil)
-}
-
-func (s *SnapTestSuite) TestWrongType(c *C) {
-	c.Check(verifyPlugYaml(&plugYaml{
-		Interface: "some-interface",
-	}), ErrorMatches, ".*can not use interface.* only `old-security` supported")
-}
-
-func (s *SnapTestSuite) TestUsesWhitelistIllegal(c *C) {
-	c.Check(verifyPlugYaml(&plugYaml{
-		Interface: "old-security",
-		SecurityDefinitions: SecurityDefinitions{
-			SecurityTemplate: "x\n"},
-	}), ErrorMatches, ".*contains illegal.*")
-	c.Check(verifyPlugYaml(&plugYaml{
-		Interface: "old-security",
-		SecurityDefinitions: SecurityDefinitions{
-			SecurityPolicy: &SecurityPolicyDefinition{
-				AppArmor: "x\n"},
-		},
-	}), ErrorMatches, ".*contains illegal.*")
-}
-
-func (s *SnapTestSuite) TestRemovePackageServiceKills(c *C) {
-	c.Skip("needs porting to new squashfs based snap activation!")
-
-	// make Stop not work
-	var sysdLog [][]string
-	systemd.SystemctlCmd = func(cmd ...string) ([]byte, error) {
-		sysdLog = append(sysdLog, cmd)
-		return []byte("ActiveState=active\n"), nil
-	}
-	yamlFile, err := makeInstalledMockSnap(s.tempdir, `name: wat
-version: 42
-apps:
- wat:
-   command: wat
-   stop-timeout: 25
-   daemon: forking
-`)
-	c.Assert(err, IsNil)
-	m, err := parseSnapYamlFile(yamlFile)
-	c.Assert(err, IsNil)
-	inter := &MockProgressMeter{}
-	c.Check(removePackageServices(m, filepath.Dir(filepath.Dir(yamlFile)), inter), IsNil)
-	c.Assert(len(inter.notified) > 0, Equals, true)
-	c.Check(inter.notified[len(inter.notified)-1], Equals, "wat_wat_42.service refused to stop, killing.")
-	c.Assert(len(sysdLog) >= 3, Equals, true)
-	sd1 := sysdLog[len(sysdLog)-3]
-	sd2 := sysdLog[len(sysdLog)-2]
-	c.Check(sd1, DeepEquals, []string{"kill", "wat_wat_42.service", "-s", "TERM"})
-	c.Check(sd2, DeepEquals, []string{"kill", "wat_wat_42.service", "-s", "KILL"})
-}
-
-func (s *SnapTestSuite) TestCopySnapDataDirectoryError(c *C) {
-	oldPath := c.MkDir()
-	newPath := "/nonono-i-can-not-write-here"
-	err := copySnapDataDirectory(oldPath, newPath)
-	c.Assert(err, DeepEquals, &ErrDataCopyFailed{
-		OldPath:  oldPath,
-		NewPath:  newPath,
-		ExitCode: 1,
-	})
-}
-
-func (s *SnapTestSuite) TestSnappyGenerateSnapSocket(c *C) {
-	service := &AppYaml{Name: "xkcd-webserver",
-		Command:      "bin/foo start",
-		Description:  "meep",
-		Socket:       true,
-		ListenStream: "/var/run/docker.sock",
-		SocketMode:   "0660",
-		Daemon:       "simple",
-	}
-	pkgPath := "/snaps/xkcd-webserver.canonical/0.3.4/"
-	aaProfile := "xkcd-webserver.canonical_xkcd-webserver_0.3.4"
-	m := snapYaml{
-		Name:    "xkcd-webserver",
-		Version: "0.3.4"}
-
-	content, err := generateSnapSocketFile(service, pkgPath, aaProfile, &m)
-	c.Assert(err, IsNil)
-	c.Assert(content, Equals, `[Unit]
-Description= Socket Unit File
-PartOf=xkcd-webserver_xkcd-webserver_0.3.4.service
-X-Snappy=yes
-
-[Socket]
-ListenStream=/var/run/docker.sock
-SocketMode=0660
-
-[Install]
-WantedBy=sockets.target
-`)
-}
-
-func (s *SnapTestSuite) TestSnappyGenerateSnapServiceWithSocket(c *C) {
-	service := &AppYaml{
-		Name:        "xkcd-webserver",
-		Command:     "bin/foo start",
-		Stop:        "bin/foo stop",
-		PostStop:    "bin/foo post-stop",
-		StopTimeout: timeout.DefaultTimeout,
-		Description: "A fun webserver",
-		Socket:      true,
-		Daemon:      "simple",
-	}
-	pkgPath := "/snaps/xkcd-webserver/0.3.4/"
-	aaProfile := "xkcd-webserver_xkcd-webserver_0.3.4"
-	m := snapYaml{Name: "xkcd-webserver",
-		Version: "0.3.4"}
-
-	generatedWrapper, err := generateSnapServicesFile(service, pkgPath, aaProfile, &m)
-	c.Assert(err, IsNil)
-	c.Assert(generatedWrapper, Equals, expectedSocketUsingWrapper)
-}
-
-func (s *SnapTestSuite) TestGenerateSnapSocketFile(c *C) {
-	srv := &AppYaml{}
-	baseDir := "/base/dir"
-	aaProfile := "pkg_app_1.0"
-	m := &snapYaml{}
-
-	// no socket mode means 0660
-	content, err := generateSnapSocketFile(srv, baseDir, aaProfile, m)
-	c.Assert(err, IsNil)
-	c.Assert(content, Matches, "(?ms).*SocketMode=0660")
-
-	// SocketMode itself is honored
-	srv.SocketMode = "0600"
-	content, err = generateSnapSocketFile(srv, baseDir, aaProfile, m)
-	c.Assert(err, IsNil)
-	c.Assert(content, Matches, "(?ms).*SocketMode=0600")
 
 }
